# Conflicts:
#	scala/scala-impl/src/org/jetbrains/plugins/scala/extensions/package.scala
#	scala/scala-impl/src/org/jetbrains/plugins/scala/lang/psi/implicits/GlobalImplicitInstance.scala

+++ scala/scala-impl/src/org/jetbrains/plugins/scala/lang/parser/parsing/patterns/Pattern1.scala
@@ -4,62 +4,40 @@ package parser
 package parsing
 package patterns
 
-import org.jetbrains.plugins.scala.lang.lexer.ScalaTokenTypes
+import org.jetbrains.plugins.scala.lang.lexer.ScalaTokenTypes.tCOLON
 import org.jetbrains.plugins.scala.lang.parser.parsing.builder.ScalaPsiBuilder
 
 /**
-* @author Alexander Podkhalyuzin
-* Date: 28.02.2008
-*/
-
-/*
- * Pattern1 ::= varid ':' TypePat
- *            | '_' ':' TypePat
- *            | Pattern2
+ * Scala 2
+ * [[Pattern1]] ::= [[PatVar]] ':' [[TypePattern]]
+ * | [[Pattern2]]
+ *
+ * Scala 3
+ * [[Pattern1]] ::= [[PatVar]] [':' [[RefinedType]] ]
+ * | 'given' [[PatVar]] ':' [[RefinedType]]
+ *
+ * @author Alexander Podkhalyuzin
+ *         Date: 28.02.2008
  */
-object Pattern1 {
+object Pattern1 extends ParsingRule {
 
-  //TODO: refactor!
-  def parse(builder: ScalaPsiBuilder): Boolean = {
-    val pattern1Marker = builder.mark
-    val backupMarker = builder.mark
-    builder.getTokenType match {
-      case ScalaTokenTypes.tIDENTIFIER =>
-        if (builder.isIdBinding) {
-          builder.advanceLexer() //Ate id
-          builder.getTokenType match {
-            case ScalaTokenTypes.tCOLON =>
-              builder.advanceLexer() //Ate :
-              backupMarker.drop()
-              if (!TypePattern.parse(builder)) {
-                builder error ScalaBundle.message("wrong.type")
-              }
-              pattern1Marker.done(ScalaElementType.TYPED_PATTERN)
-              return true
-            case _ =>
-              backupMarker.rollbackTo()
+  override def apply()(implicit builder: ScalaPsiBuilder): Boolean = {
+    val pattern1Marker = builder.mark()
+    if (PatVar()) {
+      builder.getTokenType match {
+        case `tCOLON` =>
+          builder.advanceLexer() //Ate :
+          if (!TypePattern.parse(builder)) {
+            builder.error(ScalaBundle.message("wrong.type"))
           }
-        } else {
-          backupMarker.rollbackTo()
-        }
-      case ScalaTokenTypes.tUNDER =>
-        builder.advanceLexer() //Ate _
-        builder.getTokenType match {
-          case ScalaTokenTypes.tCOLON =>
-            builder.advanceLexer() //Ate :
-            backupMarker.drop()
-            if (!TypePattern.parse(builder)) {
-              builder error ScalaBundle.message("wrong.type")
-            }
-            pattern1Marker.done(ScalaElementType.TYPED_PATTERN)
-            return true
-          case _ =>
-            backupMarker.rollbackTo()
-        }
-      case _ =>
-        backupMarker.drop()
+          pattern1Marker.done(ScalaElementType.TYPED_PATTERN)
+          return true
+        case _ =>
+          pattern1Marker.rollbackTo()
+      }
+    } else {
+      pattern1Marker.drop()
     }
-    pattern1Marker.drop()
     Pattern2.parse(builder, forDef = false)
   }
 }
